# api/api.yaml
openapi: 3.0.0
info:
  title: ZidiBackend API
  version: 1.0.0
  description: API for managing organizations and users in ZidiBackend
paths:
...
  /organizations/{organization_id}:
    get:
      summary: Get an organization by ID
      operationId: GetOrganizationById
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update an organization
      operationId: UpdateOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Delete an organization
      operationId: DeleteOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Organization deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
...
  /campaigns/{id}:
    get:
      summary: Get a campaign by ID
      operationId: getCampaignById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Campaign not found

    put:
      summary: Update a campaign by ID
      operationId: updateCampaign
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Campaign"
      responses:
        "200":
          description: Campaign updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Campaign not found

    delete:
      summary: Delete a campaign by ID
      operationId: deleteCampaign
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Campaign deleted successfully
        "404":
          description: Campaign not found
...
  /campaigns/{id}/generate-tokens:
    post:
      summary: Generate tokens for a campaign
      operationId: generateTokens
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tokens generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaignId:
                    type: string
                    format: uuid
                  tokens:
                    type: array
                    items:
                      type: string
        "404":
          description: Campaign not found
        "400":
          description: Campaign is not active
        "500":
          description: Internal server error
...
  /payments:
    post:
      summary: Create a new payment
      operationId: createPayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
      responses:
        "201":
          description: Payment successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Invalid request data
        "500":
          description: Internal server error

    get:
      summary: Get all payments
      operationId: getAllPayments
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            example: 0
      responses:
        "200":
          description: A list of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  payments:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of payments in the database.
                      payments:
                        type: array
                        items:
                          $ref: "#/components/schemas/Payment"
        "500":
          description: Internal server error
...
  /balances:
    get:
      summary: Get all balances
      operationId: getAllBalances
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"
        "401":
          description: Unauthorized

    post:
      summary: Create a new balance
      operationId: createBalance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBalanceRequest"
      responses:
        "201":
          description: Balance created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but helps with documentation
  schemas:
    Organization:
      type: object
      required:
        - id
        - email
        - contact_person_name
        - contact_person_phone
        - address
        - industry
        - company_size
        - company_name
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        contact_person_name:
          type: string
        contact_person_phone:
          type: string
        address:
          type: string
        industry:
          type: string
        company_size:
          type: integer
        company_name:
          type: string
    Campaign:
      type: object
      required:
        - id
        - campaignName
        - couponId
        - characterType
        - couponLength
        - couponNumber
        - welcomeMessage
        - questionNumber
        - amount
        - price
        - status
        - organizationId
        - communicationChannels
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        campaignName:
          type: string
          example: "Summer Promo"
        couponId:
          type: string
          example: "Alphanumeric"
        characterType:
          type: string
          example: "Alphanumeric"
        couponLength:
          type: integer
          example: 10
        couponNumber:
          type: integer
          example: 100
        welcomeMessage:
          type: string
          example: "Welcome to our summer campaign!"
        questionNumber:
          type: integer
          example: 5
        amount:
          type: number
          format: float
          example: 100.50
        price:
          type: number
          format: float
          example: 100.50
        status:
          type: string
          example: "active"
        organizationId:
          type: string
          format: uuid
          example: "d63e4c28-f8d2-4c39-b3bb-92cd5b998a21"
        communicationChannels:
          type: string
          example: "whatsapp,telegram"
          description: "Comma-separated list of communication channels: whatsapp, telegram, or both"
        createdAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-03-04T12:34:56Z"
...
    Payment:
      type: object
      required:
        - id
        - organization_id
        - campaign_id
        - amount
        - transaction_ref
        - transaction_id
        - status
      properties:
        id:
          type: string
          format: uuid
          example: "5c2a4d94-27f8-4fbd-8f74-8dd8c7c8e49c"
        organization_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        campaign_id:
          type: string
          format: uuid
          example: "2d6f2952-9a5b-4663-91e1-67d573a3f940"
        amount:
          type: number
          format: float
          example: 100.50
        transaction_ref:
          type: string
          example: "TXN123456789"
        transaction_id:
          type: string
          example: "543456789654789654"
        status:
          type: string
          enum: [pending, completed, failed]
          example: "pending"
...
    Balance:
      type: object
      required:
        - id
        - campaignId
        - startingBalance
        - amount
      properties:
        id:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        startingBalance:
          type: number
          format: float
        amount:
          type: number
          format: float
        campaignName:
          type: string
          description: "Name of the campaign linked to this balance"
    CreateBalanceRequest:
      type: object
      required:
        - campaignId
        - startingBalance
        - amount
      properties:
        campaignId:
          type: string
          format: uuid
        startingBalance:
          type: number
          format: float
        amount:
          type: number
          format: float
...
    Error:
      type: object
      required:
        - message
        - error_code
      properties:
        error_code:
          type: string
        message:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - BearerAuth: []